Rename space.js to spaces.js
