+++ lib/spaces.js
