+++ lib/sub-commands/copy-fields.js
@@ -1,6 +1,6 @@
 const log = require('npmlog')
 const _ = require('lodash')
-const getSpace = require('./../space')
+const {getSpace} = require('./../spaces')
 const {
   addFields, getFields, getSource, getTargets, handleContentTypesRejection
 } = require('./../content-types')
