+++ lib/sub-commands/delete-fields.js
@@ -1,6 +1,6 @@
 const log = require('npmlog')
 const _ = require('lodash')
-const getSpace = require('./../space')
+const {getSpace} = require('./../spaces')
 const {
   getTargets, omitFields, deleteFields: delFields, handleContentTypesRejection
 } = require('./../content-types')
